fix: ruff issues
